move parsing out
